+++ app/actions/wallet-settings-actions.js
@@ -1,6 +1,13 @@
 // @flow
 import Action from './lib/Action';
 
+export type WalletExportToFileParams = {
+  walletId: string,
+  exportType: string,
+  filePath: string,
+  password: ?string
+};
+
 export default class WalletSettingsActions {
   cancelEditingWalletField: Action<any> = new Action();
   startEditingWalletField: Action<{ field: string }> = new Action();
@@ -8,4 +15,5 @@ export default class WalletSettingsActions {
   updateWalletField: Action<{ field: string, value: string }> = new Action();
   // eslint-disable-next-line max-len
   updateWalletPassword: Action<{ walletId: string, oldPassword: ?string, newPassword: ?string }> = new Action();
+  exportToFile: Action<WalletExportToFileParams> = new Action();
 }