+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Changelog
 - Opt-in mode for sending logs to the remote server
 - Added support for Cmd+H hotkey shortcut for hiding application window on OSX ([PR 404](https://github.com/input-output-hk/daedalus/pull/404))
 - Setup environment variable for testnet/mainnet mode ([PR 400](https://github.com/input-output-hk/daedalus/pull/400))
+- Export wallet to file ([PR 426](https://github.com/input-output-hk/daedalus/pull/426))
 - Add theming options in settings ([PR 370](https://github.com/input-output-hk/daedalus/pull/398))
 - About page, available from the system menu ([PR 430](https://github.com/input-output-hk/daedalus/pull/430))
 
@@ -96,6 +97,7 @@ Changelog
 - Fixed failing acceptance tests ([PR 424](https://github.com/input-output-hk/daedalus/pull/424))
 - Show correct error message on sending money to Ada redemption address ([PR 423](https://github.com/input-output-hk/daedalus/pull/423))
 - Fixed ENOENT error on log file rotation ([PR 428](https://github.com/input-output-hk/daedalus/pull/428))
+- Fixed logging error ([PR 431](https://github.com/input-output-hk/daedalus/pull/431))
 
 ### Chores